refactor: extract save request to local storage function into separate file
