+++ public/locales/ru/lang.json
@@ -63,6 +63,7 @@
     "key": "Ключ",
     "value": "Значение",
     "addHeader": "Добавить заголовок",
+
     "schema": {
       "schemaHeader": "Схема GraphQL",
       "schema": "Схема",
@@ -71,9 +72,9 @@
       "missing": "Неверный ответ интроспекции: отсутствует __schema"
     },
     "header": {
-      "headers": "Заголовки:",  
-      "key": "Ключ",   
-      "value": "Значение",   
+      "headers": "Заголовки:",
+      "key": "Ключ",
+      "value": "Значение",
       "addHeader": "Добавить заголовок",
       "Headers": "Заголовки",
       "variables": "Переменные",
